requirements.txt
asgiref==3.7.2
boto3==1.21.3
botocore==1.24.46
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.1.1
django-crispy-forms==1.14.0
django-storages==1.12.3
django-tables2==2.4.1
docker==6.1.3
idna==3.4
jmespath==0.10.0
packaging==23.2
Pillow==9.2.0
psycopg2-binary==2.9.3
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.4
urllib3==1.26.18
websocket-client==1.6.4
